@@ -1,7 +1,8 @@
 """Getswish python client library"""
+
 from .client import SwishClient  # noqa F401
 from .environments import Certificate, Certificates, Environment, ProductionEnvironment, TestEnvironment  # noqa F401
 from .exceptions import SwishError  # noqa F401
 from .models import Payment, Payout, Refund  # noqa F401
 
-__version__ = "0.2.4"
+__version__ = "0.3.0"
