@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field
 from pathlib import Path
 
-cert_base = Path(__file__).parent.parent.parent.resolve() / "mss_test_1.9" / "Getswish_Test_Certificates"
+cert_base = Path(__file__).parent.parent.parent.resolve() / "mss_test_2.0" / "Getswish_Test_Certificates"
 
 
 @dataclass
@@ -32,9 +32,7 @@ TestCertificates = Certificates(
         public=f"{cert_base}/Swish_Merchant_TestCertificate_1234679304.pem",
         private_key=f"{cert_base}/Swish_Merchant_TestCertificate_1234679304.key",
     ),
-    Certificate(
-        public=f"{cert_base}/Swish_TLS_RootCA.pem"
-    ),
+    Certificate(public=f"{cert_base}/Swish_TLS_RootCA.pem"),
     Certificate(
         public=f"{cert_base}/Swish_Merchant_TestSigningCertificate_1234679304.pem",
         private_key=f"{cert_base}/Swish_Merchant_TestSigningCertificate_1234679304.key",
